공고글 작성하기 기능 및 시큐리티 수정

+++ src/main/java/com/kakaoseventeen/dogwalking/dog/DogJpaRepository.java
@@ -3,7 +3,7 @@ package com.kakaoseventeen.dogwalking.dog;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface DogJpaRepository extends JpaRepository<Dog, Integer> {
+public interface DogJpaRepository extends JpaRepository<Dog, Long> {
 
-    List<Dog> findDogsByMemberId(int memberId);
+    List<Dog> findDogsByMemberId(Long memberId);
 }
